\name{batman-package}
\alias{batman-package}
\docType{package}
\title{
Bayesian AuTomated Metabolite Analyser for NMR spectra (BATMAN)
}

\description{
BATMAN deconvolves resonance peaks from NMR spectra of complex mixtures and obtains 
concentration estimates for the corresponding metabolites automatically.
This is achieved through a database of spectral profiles for known metabolites and 
a Bayesian Markov Chain Monte Carlo algorithm. Users have the options to  
specify the multiplet ppm position, position shift range, peak width range and so on.
Parallel processing is available if processing several spectra.
\tabular{ll}{\tab
\if{html}{\figure{batmanlogo.png}}
\if{latex}{\figure{batmanlogo.png}{options: width=1.5in}} \cr
}
}
\details{	
\tabular{ll}{
Package: \tab batman\cr
Type: \tab Package\cr
Version: \tab 1.2.0\cr
Date: \tab 2013-08-06\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}}
\author{
 Jie Hao <j.hao@ic.ac.uk>, William Astle <william.astle@mail.mcgill.ca>,   
 Maria De Iorio <m.deiorio@ucl.ac.uk>, Timothy Ebbels <t.ebbels@ic.ac.uk>
}
\references{
Hao, J., et al., BATMAN--an R package for the automated quantification of metabolites from nuclear magnetic resonance spectra using a Bayesian model. Bioinformatics, 2012. 28(15): p. 2088-90.

http://bioinformatics.oxfordjournals.org/content/28/15/2088



Astle, W., et al., A Bayesian Model of NMR Spectra for the Deconvolution and Quantification of Metabolites in Complex Biological Mixtures. Journal of the American Statistical Association, 2012. 107(500): p. 1259-1271.

http://www.tandfonline.com/doi/abs/10.1080/01621459.2012.695661#.UgEf-hbZa4k
}
\keyword{ package }

