% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/batman.R
\docType{package}
\name{batman}
\alias{batman}
\alias{batman-package}
\title{Convert categorical representations of logicals to actual logicals}
\description{
Survey systems and other third-party data sources commonly use non-standard
representations of logical values when it comes to qualitative data - "Yes", "No" and "N/A", say.
batman is a package designed to seamlessly convert these into actual logical values.
}
\seealso{
\code{to_logical}
}

