% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/languages.R
\docType{data}
\name{categorical_booleans}
\alias{categorical_booleans}
\title{TRUE/FALSE equivalents in categorical data for various
languages}
\format{A data.frame of three columns:
\describe{
  \item{language}{a two-letter language code vector of equivalents to TRUE;}
  \item{cat}{the actual categorical variable in that language;}
  \item{value}{the value (TRUE or FALSE) that \code{cat} represents.}
}}
\usage{
categorical_booleans
}
\description{
A dataset containing the equivalents of TRUE or FALSE in
categorical or user-submitted data, localised to various languages
}
\seealso{
\code{\link{to_logical}}, which uses this dataset, and
\code{\link{get_languages}} to see what languages are available.
}
\keyword{datasets}

