% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/languages.R
\name{get_languages}
\alias{get_languages}
\title{Get language codes for batman-supported languages}
\usage{
get_languages()
}
\description{
retrieves a list of language codes for languages
supported by the \code{language} parameter in \code{\link{to_logical}}.
}
\examples{
get_languages()
# [1] "en"
}
\seealso{
\code{\link{categorical_booleans}}, the underlying dataset,
or \code{\link{to_logical}}, which uses that dataset.
}

