% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/languages.R
\docType{data}
\name{categorical_booleans}
\alias{categorical_booleans}
\title{TRUE/FALSE equivalents in categorical data for various
languages}
\format{A list of named lists, each one containing
two columns:
\describe{
  \item{true}{a character vector of equivalents to TRUE}
  \item{false}{a character vector of equivalents to FALSE}
}}
\usage{
categorical_booleans
}
\description{
A dataset containing the equivalents of TRUE or FALSE in
categorical or user-submitted data, localised to various languages
}
\seealso{
\code{\link{to_logical}}, which uses this dataset, and
\code{\link{get_languages}} to see what languages are available.
}
\keyword{datasets}

